[PugiXml] Fix serialization int64 types on Linux

+++ tests/bitserializer_cpprest_json_tests/cpprest_json_archive_tests.cpp
@@ -82,7 +82,7 @@ TEST(JsonRestCpp, SerializeTwoDimensionalArray) {
 
 TEST(JsonRestCpp, ShouldLoadToArrayWithLesserAmountOfElements) {
 	TestSerializeArray<JsonArchive, bool, 7, 5>();
-	TestSerializeArray<JsonArchive, int64_t, 7, 5>();
+	TestSerializeArray<JsonArchive, int, 7, 5>();
 	TestSerializeArray<JsonArchive, double, 7, 5>();
 	TestSerializeArray<JsonArchive, std::string, 7, 5>();
 	TestSerializeArray<JsonArchive, TestPointClass, 7, 5>();
@@ -90,7 +90,7 @@ TEST(JsonRestCpp, ShouldLoadToArrayWithLesserAmountOfElements) {
 
 TEST(JsonRestCpp, ShouldLoadToArrayWithBiggerAmountOfElements) {
 	TestSerializeArray<JsonArchive, bool, 5, 7>();
-	TestSerializeArray<JsonArchive, int64_t, 5, 7>();
+	TestSerializeArray<JsonArchive, int, 5, 7>();
 	TestSerializeArray<JsonArchive, double, 5, 7>();
 	TestSerializeArray<JsonArchive, std::string, 5, 7>();
 	TestSerializeArray<JsonArchive, TestPointClass, 5, 7>();
