+++ tests/bitserializer_rapidjson_tests/rapidjson_archive_tests.cpp
@@ -85,7 +85,7 @@ TEST(RapidJsonArchive, SerializeTwoDimensionalArray) {
 
 TEST(RapidJsonArchive, ShouldLoadToArrayWithLesserAmountOfElements) {
 	TestSerializeArray<JsonArchive, bool, 7, 5>();
-	TestSerializeArray<JsonArchive, int64_t, 7, 5>();
+	TestSerializeArray<JsonArchive, int, 7, 5>();
 	TestSerializeArray<JsonArchive, double, 7, 5>();
 	TestSerializeArray<JsonArchive, std::string, 7, 5>();
 	TestSerializeArray<JsonArchive, TestPointClass, 7, 5>();
@@ -93,7 +93,7 @@ TEST(RapidJsonArchive, ShouldLoadToArrayWithLesserAmountOfElements) {
 
 TEST(RapidJsonArchive, ShouldLoadToArrayWithBiggerAmountOfElements) {
 	TestSerializeArray<JsonArchive, bool, 5, 7>();
-	TestSerializeArray<JsonArchive, int64_t, 5, 7>();
+	TestSerializeArray<JsonArchive, int, 5, 7>();
 	TestSerializeArray<JsonArchive, double, 5, 7>();
 	TestSerializeArray<JsonArchive, std::string, 5, 7>();
 	TestSerializeArray<JsonArchive, TestPointClass, 5, 7>();
